fix(fx): artefactos del pipeline de efectos y carga de modelos
Tanda de correcciones de la misma familia de artefactos de Hex-Rays:

- Model_Misc / Model_Players: 17 bucles cuya condicion tomaba la base del
  ARGUMENTO en vez del inicio del bucle (IDA escribe `v0 = 197; do {... v0-196
  ...} while (v0 - 197 < 2)`).  Era la causa de que Inferno no cargara
  Stone02.bmd.

- MoveEffect / Effect_Create / Particle_Spawn: campos DWORD del slot leidos
  como float.  El contador (+0x18), el SubType (+4) y el puntero al owner
  (+252) se convertian numericamente en vez de reinterpretarse, asi que ramas
  enteras nunca corrian y el owner quedaba en NULL.  Tambien los literales
  -NAN reemplazados por los enteros de IDA y la direccion del Aqua Beam, que
  arrancaba con un valor sobrante en in1[0]/in1[2].

- Cursor_Render y otros ocho sitios: `(int)WorldTime` satura pasados 24.8 dias
  de uptime (2^31 ms); IDA usa `(__int64)`.  Congelaba cursor, agua y UV
  animadas.

- Render_WorldHelpers/RenderBoids, Weather worlds 7 y 11-16, y el gate nulo del
  pool de boids.

- Monster / Entity_UpdateRender / RenderLinkObject / Render_Frame: limpieza de
  la instrumentacion usada para la caza y ajustes menores verificados contra
  IDA.

+++ src/Input/Cursor_Render.cpp
@@ -35,7 +35,14 @@ void __cdecl Cursor_Render(void) {
 
     int u_bits = 0, v_bits = 0;
     // Frame = (int64)(WorldTime * 0.01) % 6  — IDA lo emite con __int64 cast explícito.
-    int frame = (int)((long long)((double)(int)DAT_05826e08 * 0.0099999998)) % 6;
+    // 2026-09-03 -- ANIMACION DEL CURSOR CONGELADA (y de todo lo que depende
+    // de WorldTime).  IDA: `Frame = (int)(__int64)(WorldTime * 0.0099999998) % 6;`
+    // -- multiplica el FLOAT y recien despues convierte a __int64.  El port
+    // castea a `int` ANTES, y WorldTime = timeGetTime() pasa de 2^31 ms a las
+    // ~24.8 dias de uptime de la maquina: el cast satura y el frame queda
+    // clavado.  Por eso el cursor sobre NPC se ve estatico en una maquina con
+    // mucho uptime y normal en una recien reiniciada.  Habia 9 sitios iguales.
+    int frame = (int)(long long)((double)DAT_05826e08 * 0.0099999998) % 6;
     if (frame == 1 || frame == 3 || frame == 5) u_bits = 0x3F000000;  // 0.5f
     if (frame == 2 || frame == 3 || frame == 4) v_bits = 0x3F000000;  // 0.5f
 
@@ -98,7 +105,7 @@ void __cdecl Cursor_Render(void) {
         return;
     }
     if (DAT_07eaa134 == 2) {
-        float10 fv = fsin((float10)(int)DAT_05826e08 * (float10)_DAT_00552914);
+        float10 fv = fsin((float10)(long long)DAT_05826e08 * (float10)_DAT_00552914);
         if (fv <= (float10)FloatZero) {
             FUN_005126e0(7, (float)(int)DAT_083a427c + _DAT_00552660,
                             (float)(int)DAT_083a4278 + _DAT_005529fc,
